+++ gh-page/gh-page.js
@@ -1,6 +1,7 @@
 "use strict";
 (async () => {
     wikiparse.setConfig(await (await fetch('/wikiparser-node/config/default.json')).json());
+    localStorage.removeItem('codemirror-mediawiki-addons');
     const container = document.querySelector('#container'), languages = [...document.querySelectorAll('input[name="language"]')], editor = (await monaco).editor.create(container, {
         automaticLayout: true,
         theme: 'monokai',